refactor(EditableSharingModal): Colocate with his parent
