+++ packages/cozy-sharing/src/components/ShareModal/ShareModal.jsx
@@ -2,10 +2,10 @@ import omit from 'lodash/omit'
 import React from 'react'
 import { useLocation } from 'react-router'
 
+import EditableSharingModal from './EditableSharingModal'
 import { SharingDetailsModal } from './SharingDetailsModal'
 import { useSharingContext } from '../../hooks/useSharingContext'
 import withLocales from '../../withLocales'
-import EditableSharingModal from '../EditableSharingModal'
 
 export const ShareModal = withLocales(props => {
   const location = useLocation?.()
